Implemented Product CRUD. Changed logic for Groups

+++ labas/src/daxzel/model/services/impl/ProductServiceImpl.java
@@ -1,24 +1,33 @@
 package daxzel.model.services.impl;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.List;
 
-import daxzel.model.DAO.ProductDAO;
 import daxzel.model.domains.Product;
+import daxzel.model.domains.Group;
 import daxzel.model.services.ProductService;
+import daxzel.model.DAO.ProductDAO;
+import daxzel.model.DAO.GroupDAO;
 
 @Service
 public class ProductServiceImpl implements ProductService {
 
 	@Autowired
 	private ProductDAO productDAO;
+	
+	@Autowired
+	private GroupDAO groupDAO;
 
 	@Transactional
 	public void add(Product entity)
 	{
+		if (entity.getGroup()==null)
+		{
+			Group group = new Group();
+			entity.setGroup(group);
+		}
 		productDAO.addOrUpdate(entity);
 	}
 
@@ -37,7 +46,34 @@ public class ProductServiceImpl implements ProductService {
 	@Transactional
 	public void remove(Long id)
     {
-		productDAO.remove(id);
+		Product product = productDAO.getByID(id);
+		if (product!=null)
+		{
+			Group group = product.getGroup();
+			if (group.getProducts().size()==1)
+			{
+				groupDAO.remove(group.getKey());
+			}			
+			productDAO.remove(id);
+		}
+    }
+	
+	@Transactional
+	public Product findProductByNCP(Long NCP)
+	{
+		return productDAO.findProductByNCP(NCP);
+	}
+	
+	@Transactional
+	public void removeProductByNCP(Long NCP)
+	{
+		productDAO.removeProductByNCP(NCP);
+	}
+	
+	@Transactional
+	public List<Group> getAllGroups()
+    {
+		return groupDAO.getAll();
     }
 	
 }
